requirements.txt
flask==3.1.0
pillow==11.1.0
pytesseract==0.3.13
opencv-python==4.11.0.86
numpy==2.2.3
python-dotenv==1.0.1
gunicorn==21.2.0
